When wrapping a c array in a numpy, allow dtype to be set if required

+++ cfficloak.py
@@ -1088,10 +1088,10 @@ class nparray(object):
     :param cffi.CData cdata: array object, expected to be uint8_t or equivalent
     :return: wrapped numpy array object
     """
-    def __init__(self, _cdata, size = -1):
+    def __init__(self, _cdata, size=-1, dtype=numpy.uint8):
         self.__cdata = _cdata
         self.__buff = _global_ffi.buffer(_cdata, size=size)
-        self.__nparray = numpy.frombuffer(self.__buff, dtype=numpy.uint8)
+        self.__nparray = numpy.frombuffer(self.__buff, dtype=dtype)
 
     def __getattr__(self, item):
         return getattr(self.__nparray, item)
